content-security-policy/script-src/script-src-strict_dynamic_double_policy_honor_whitelist.html

<!DOCTYPE HTML>
<html>

<head>
    <title>Whitelists in a separate policy are honored with `strict-dynamic` in the script-src directive.</title>
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>
    <script src='../support/checkSecurityPolicyViolations.sub.js' nonce='dummy'></script>
    <!-- CSP served:
    1) Content-Security-Policy: script-src 'strict-dynamic' 'nonce-dummy'
    2) Content-Security-Policy: script-src 'self' 'nonce-dummy'
    -->
</head>

<body>
    <h1>Whitelists in a separate policy are honored with `strict-dynamic` in the script-src directive.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        async_test(function(t) {
            var e = document.createElement('script');
            e.src = 'http://example.org/script.js';
            e.onload = t.unreached_func('OnLoad should not be triggered.');
            document.body.appendChild(e);
            t.done();
        }, 'Non-whitelisted script injected via `appendChild` is not allowed with `strict-dynamic` + a nonce+whitelist double policy.');
    </script>

    <script nonce='dummy'>
        async_test(function(t) {
            setTimeout(t.step_func_done(function() {
                assert_violations({
                    'script-src': 1 // One violation from the second policy.
                });
                assert_original_policies(["script-src 'self' 'nonce-dummy'"]);
            }), 1000);
        }, 'All the expected CSP violation reports have been fired.');
    </script>

</body>

</html>
